Mon, Jul 17, 2023, 1:34 PM +03:00

+++ volume/volume.c
@@ -54,7 +54,7 @@ int main(int argc, char *argv[])
     while(fread(&buffer, sizeof(uint16_t), 1, input))
     {
         printf("%i\n", buffer);
-        buffer = buffer * factor;
+        buffer = (float)buffer * factor;
         printf("%i\n", buffer);
         fwrite(&buffer, sizeof(uint16_t), 1, output);
     }
